# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/weather": {
      "get": {
        "description": "기상청 자외선 + 에어코리아 대기오염 결합 스냅샷. lat/lon이 있으면 근접 지역, 없으면 기본 지역. 각 지표는 API 실패 시 null(측정 불가)로 응답한다.",
        "operationId": "WeatherController_getCurrentWeather",
        "parameters": [
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "description": "위도",
            "schema": {
              "example": 37.5665,
              "type": "number"
            }
          },
          {
            "name": "lon",
            "required": false,
            "in": "query",
            "description": "경도",
            "schema": {
              "example": 126.978,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherSnapshotDto"
                }
              }
            }
          }
        },
        "summary": "현재 날씨 스냅샷 (자외선 + 대기오염)",
        "tags": [
          "weather"
        ]
      }
    },
...
    "/care/weather": {
      "get": {
        "description": "좌표만 받아 서버가 오늘 날씨를 직접 조회하고 케어 루틴+실제 제품을 즉시 반환한다 (source: CACHED | FALLBACK | LIVE). CACHED/FALLBACK이면 jobId로 GET /jobs/:id polling해 LIVE로 교체한다.",
        "operationId": "CareController_weather",
        "parameters": [
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "description": "위도 (-90 ~ 90)",
            "schema": {
              "example": 37.5665,
              "type": "number"
            }
          },
          {
            "name": "lon",
            "required": false,
            "in": "query",
            "description": "경도 (-180 ~ 180)",
            "schema": {
              "example": 126.978,
              "type": "number"
            }
          },
          {
            "name": "refresh",
            "required": false,
            "in": "query",
            "description": "직전 결과를 무시하고 새로 생성",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarePlanFastResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "날씨 기반 케어 루틴+제품 빠른 경로",
        "tags": [
          "care"
        ]
      }
    },
    "/care/skin": {
      "post": {
        "description": "지정한 진단의 피부 측정값을 기준으로 케어 루틴+실제 제품을 생성한다.",
        "operationId": "CareController_skin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CareDiagnosisRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarePlanFastResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "피부 상태 기반 케어 루틴+제품 빠른 경로",
        "tags": [
          "care"
        ]
      }
    },
    "/care/combined": {
      "post": {
        "description": "지정한 진단의 피부 측정값과 그 진단에 연결된 날씨(diagnosis.weatherSnapshotId)를 함께 반영한다.",
        "operationId": "CareController_combined",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CareDiagnosisRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarePlanFastResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "날씨+피부 상태 복합 케어 루틴+제품 빠른 경로",
        "tags": [
          "care"
        ]
      }
    },
...
  },
  "info": {
    "title": "Todayskin API",
    "description": "날씨 연동 AI 피부 진단 및 맞춤형 화장품 추천 서비스",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "WeatherSnapshotDto": {
        "type": "object",
        "properties": {
          "observedAt": {
            "type": "string",
            "example": "2026-08-04T03:00:00.000Z",
            "description": "관측 시각(ISO8601, UTC)"
          },
          "regionName": {
            "type": "string",
            "example": "서울특별시",
            "description": "표시용 지역명(시/도)"
          },
          "districtName": {
            "type": "string",
            "example": "해운대구",
            "description": "시/군/구 표시명 (없으면 null)",
            "nullable": true
          },
          "source": {
            "type": "string",
            "enum": [
              "LIVE",
              "CACHED",
              "UNAVAILABLE"
            ],
            "example": "LIVE",
            "description": "데이터 출처"
          },
          "uvCollectionFailed": {
            "type": "boolean",
            "description": "N42: 자외선 수집이 실패했는지. true면 값이 비어 있는 이유가 \"측정값 없음\"이 아니라 \"수집 실패\"다."
          },
          "airCollectionFailed": {
            "type": "boolean",
            "description": "N42: 대기질 수집이 실패했는지."
          },
          "nowcastCollectionFailed": {
            "type": "boolean",
            "description": "N53: 기온·습도(초단기실황) 수집이 실패했는지."
          },
          "uvIndex": {
            "type": "number",
            "description": "자외선 지수(현재 시점 예보)",
            "nullable": true
          },
          "uvStatus": {
            "type": "string",
            "enum": [
              "low",
              "moderate",
              "high",
              "veryHigh",
              "danger"
            ],
            "description": "자외선 등급 (기상청 5단계: 낮음·보통·높음·매우높음·위험)",
            "nullable": true
          },
          "uvIndexPeak": {
            "type": "number",
            "description": "오늘 남은 시간대 중 예상 자외선 최댓값",
            "nullable": true
          },
          "uvStatusPeak": {
            "type": "string",
            "enum": [
              "low",
              "moderate",
              "high",
              "veryHigh",
              "danger"
            ],
            "description": "자외선 피크 등급 (기상청 5단계)",
            "nullable": true
          },
          "uvIndexPeakHour": {
            "type": "number",
            "description": "피크 자외선이 나오는 시각(0~23시)",
            "nullable": true
          },
          "ozonePpm": {
            "type": "number",
            "description": "오존(ppm)",
            "nullable": true
          },
          "ozoneStatus": {
            "type": "string",
            "enum": [
              "good",
              "moderate",
              "bad",
              "veryBad"
            ],
            "description": "오존 등급",
            "nullable": true
          },
          "pm25": {
            "type": "number",
            "description": "초미세먼지(PM2.5)",
            "nullable": true
          },
          "pm25Status": {
            "type": "string",
            "enum": [
              "good",
              "moderate",
              "bad",
              "veryBad"
            ],
            "description": "초미세먼지 등급",
            "nullable": true
          },
          "pm10": {
            "type": "number",
            "description": "미세먼지(PM10)",
            "nullable": true
          },
          "pm10Status": {
            "type": "string",
            "enum": [
              "good",
              "moderate",
              "bad",
              "veryBad"
            ],
            "description": "미세먼지 등급",
            "nullable": true
          },
          "caiValue": {
            "type": "number",
            "description": "통합대기환경지수(CAI)",
            "nullable": true
          },
          "caiStatus": {
            "type": "string",
            "enum": [
              "good",
              "moderate",
              "bad",
              "veryBad"
            ],
            "description": "CAI 등급",
            "nullable": true
          },
          "no2Value": {
            "type": "number",
            "description": "이산화질소(NO2)",
            "nullable": true
          },
          "so2Value": {
            "type": "number",
            "description": "아황산가스(SO2)",
            "nullable": true
          },
          "coValue": {
            "type": "number",
            "description": "일산화탄소(CO)",
            "nullable": true
          },
          "temperature": {
            "type": "number",
            "description": "N53: 기온(°C, 기상청 초단기실황 T1H)",
            "nullable": true
          },
          "humidity": {
            "type": "number",
            "description": "N53: 상대습도(%, 기상청 초단기실황 REH)",
            "nullable": true
          }
        },
        "required": [
          "observedAt",
          "regionName",
          "source"
        ]
      },
...
      "CareEvidenceDto": {
        "type": "object",
        "properties": {
          "sourceName": {
            "type": "object",
            "description": "출처 기관/문서명",
            "nullable": true
          },
          "sourceUrl": {
            "type": "object",
            "description": "web_search로 확인된 실제 URL",
            "nullable": true
          },
          "sourceType": {
            "type": "string",
            "enum": [
              "WHO",
              "FDA",
              "식약처",
              "AAD",
              "PubMed",
              "없음"
            ],
            "example": "WHO"
          }
        },
        "required": [
          "sourceType"
        ]
      },
      "CareRoutineStepDto": {
        "type": "object",
        "properties": {
          "phase": {
            "type": "string",
            "description": "단계 구분 (예: 세안 후, 외출 전, 자기 전)"
          },
          "step": {
            "type": "string",
            "description": "무엇을 하는 단계인지"
          },
          "ingredient": {
            "type": "object",
            "description": "핵심 성분",
            "nullable": true
          },
          "amount": {
            "type": "object",
            "description": "바르는 양 (예: 500원 동전 크기)",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "description": "오늘 수치/피부상태 기반 이유"
          },
          "evidence": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CareEvidenceDto"
              }
            ]
          }
        },
        "required": [
          "phase",
          "step",
          "reason"
        ]
      },
      "CareProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "실제 제품명"
          },
          "url": {
            "type": "string",
            "description": "web_search로 확인된 구매 페이지 URL"
          },
          "reason": {
            "type": "string",
            "description": "이 제품을 고른 이유"
          },
          "evidence": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CareEvidenceDto"
              }
            ]
          }
        },
        "required": [
          "name",
          "url",
          "reason"
        ]
      },
      "CarePlanDto": {
        "type": "object",
        "properties": {
          "careType": {
            "type": "string",
            "enum": [
              "weather",
              "skin",
              "combined"
            ]
          },
          "routine": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CareRoutineStepDto"
            }
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CareProductDto"
            }
          },
          "medicalDisclaimer": {
            "type": "object",
            "nullable": true,
            "description": "의료 면책 문구"
          }
        },
        "required": [
          "careType",
          "routine",
          "products"
        ]
      },
      "CarePlanFastResponseDto": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string",
            "enum": [
              "CACHED",
              "FALLBACK",
              "LIVE"
            ],
            "example": "FALLBACK"
          },
          "jobId": {
            "type": "string",
            "description": "LIVE 교체용 job id"
          },
          "generatedAt": {
            "type": "string",
            "description": "이 결과가 만들어진 시각"
          },
          "plan": {
            "$ref": "#/components/schemas/CarePlanDto"
          }
        },
        "required": [
          "source",
          "plan"
        ]
      },
      "CareDiagnosisRequestDto": {
        "type": "object",
        "properties": {
          "diagnosisId": {
            "type": "string",
            "description": "이 진단 기준으로 생성"
          },
          "refresh": {
            "type": "boolean",
            "description": "직전 결과를 무시하고 새로 생성"
          }
        }
      },
...
    }
  }
}